better importer returns

+++ src/loss/categorical_crossentropy.rs
@@ -6,14 +6,14 @@ pub struct LossCategoricalCrossentropy;
 impl Loss for LossCategoricalCrossentropy {
 
     // todo: add support for y_true to be multidimensional
-    fn calculate(y_pred: &mut FynnArray, y_true: Vec<usize>) -> f64 {
+    fn calculate(y_pred: &mut FynnArray, y_true: Vec<u32>) -> f64 {
         let y_pred_clipped = MathHelpers::clip(y_pred, 0.0000001, 0.9999999);
 
         let correct_confidence: Vec<f64> = y_pred_clipped
             .iter()
             .zip(y_true.iter())
             .map(|(v, v_true)| {
-                let val = *v.get(*v_true).unwrap();
+                let val = *v.get(*v_true as usize).unwrap();
                 -val.ln()
             })
             .collect();
